extract axios logic
